// src/components/simpleGames/TreasuresAndMonsters/interfaces.ts
export type CardType = "monster" | "treasure" | "player";

export interface ICardTexts {
  title: string;
  description: string;
  quote: string;
}

export interface IBasicCard extends ICardTexts {
  id: number;
  type: CardType;
  title: string;
  description: string;
  quote: string;
}

export interface IMonsterCard extends IBasicCard {
  id: number;
  title: string;
  type: "monster";
  description: string;
  quote: string;
  attack: number;
  health: number;
}

export interface IPlayerCard extends IBasicCard {
  id: number;
  title: string;
  type: "player";
  description: string;
  quote: string;
  attack: number;
  health: number;
}

export interface ITreasureCard extends IBasicCard {
  id: number;
  title: string;
  type: "treasure";
  description: string;
  quote: string;
  bonuses: {
    attack?: number;
    health?: number;
    gold?: number;
  };
}

/*******************************************************
 * Игровые комбинации карт, образующие отдельные единицы
 *
 * комбинация может передвигаться по полю внутри фигуры
 * фигура - контейнер для комбо!
 ********************************************************/

/** комбинация игрока - постоянно в игре **/
export interface IPlayerCardCombo {
  playerCard: IPlayerCard;
  // награбленное
  treasureCards: ITreasureCard[];
  // побежденные монстры
  monstersWin: IMonsterCard[];
  // умер от монстра
  diedByMonster?: IMonsterCard;
}

/** монстр - существует до столкновения с игроком **/
export interface IMonsterCombo {
  monsterCard: IMonsterCard;
  // прячет сокровища
  treasureCards: ITreasureCard[];
  // вскрыто или нет, если не вскрыто - знак вопроса или данжен
  isOpened?: boolean;
}

/** сокровища - могут лежать пачкой **/
export interface ITreasureCombo {
  treasureCards: ITreasureCard[];
  // вскрыто или нет, если не вскрыто - знак вопроса или данжен
  isOpened?: boolean;
}

/**
 * комбинации в порядке важности трактовки
 * можно использовать как универсальный объект
 * для
 * - стартов
 * - обозначения фигур
 * **/
export interface ICombos {
  playerCombo?: IPlayerCardCombo;
  monsterCombo?: IMonsterCombo;
  treasureCombo?: ITreasureCombo;
}

/***************
 * Типы клеток поля
 ****************/
// Извлекаем тип для отдельного элемента массива terrains
export type TerrainType = "mountain" | "water" | "forest" | "sand" | "grass" | "ground";

export interface ITerrainRule {
  type: TerrainType;
  speedK: number;
  probability: number;
}

export type ITerrainRulesMap = Record<TerrainType, ITerrainRule>;

/***************
 * подготовка к старту игры
 ****************/
export interface ITerrainGenerator {
  generateCells(): ITerrainCell[][];
}

export interface ITerrainCell {
  terrainType: TerrainType;
}

export interface IMapGenerator {
  run(): IMapGeneratorResult;
}
// ...
export interface IMapGeneratorResult {
  startCells: StartCells;
  startCombos: StartCombos;
}

export interface IMapStartCell extends ICombos, ITerrainCell {
  terrainType: TerrainType;
  playerCombo?: IPlayerCardCombo;
  monsterCombo?: IMonsterCombo;
  treasureCombo?: ITreasureCombo;
}

export type StartCells = IMapStartCell[][];

export interface IComboPlace extends ICombos {
  column: number;
  row: number;
  playerCombo?: IPlayerCardCombo;
  monsterCombo?: IMonsterCombo;
  treasureCombo?: ITreasureCombo;
}

export type StartCombos = IComboPlace[];